show shortcuts only on the start of the line
